build: Updating mediawiki/mediawiki-phan-config to 0.9.0
Depends-On: I719fec91cb765a1673299974ec0eea59c37ce72f

+++ includes/specials/SpecialCentralAuth.php
@@ -361,7 +361,7 @@ class SpecialCentralAuth extends SpecialPage {
 		$globalUser = $this->mGlobalUser;
 
 		$reg = $globalUser->getRegistration();
-		$age = $this->prettyTimespan( wfTimestamp( TS_UNIX ) - wfTimestamp( TS_UNIX, $reg ) );
+		$age = $this->prettyTimespan( (int)wfTimestamp( TS_UNIX ) - (int)wfTimestamp( TS_UNIX, $reg ) );
 		$attribs = [
 			'username' => htmlspecialchars( $globalUser->getName() ),
 			'registered' => htmlspecialchars(
